Moved PartitionDataManager to separate file
